rename HtmlUtil namespace
